[Refactoring] Quick refactoring into small modules
This commit refactors the repo and creates new modules layers and config.
