+++ mistral/config/model_config.py
@@ -1,0 +1,22 @@
+from dataclasses import dataclass
+import torch
+
+@dataclass
+class ModelArgs:
+    dim: int
+    n_layers: int
+    head_dim: int
+    hidden_dim: int
+    n_heads: int
+    n_kv_heads: int
+    sliding_window: int
+    norm_eps: float
+    vocab_size: int
+
+    max_batch_size: int = 0
+
+
+def repeat_kv(keys: torch.Tensor, values: torch.Tensor, repeats: int, dim: int):
+    keys = torch.repeat_interleave(keys, repeats=repeats, dim=dim)
+    values = torch.repeat_interleave(values, repeats=repeats, dim=dim)
+    return keys, values
