+++ mistral/layers/attention.py
@@ -1,0 +1,72 @@
+from torch import Tensor
+from torch import nn
+from mistral.config.model_config import ModelArgs, repeat_kv
+from mistral.cache import CacheView
+from xformers.ops.fmha import memory_efficient_attention
+from mistral.rope import apply_rotary_emb
+
+class Attention(nn.Module):
+    def __init__(self, args: ModelArgs):
+        super().__init__()
+        self.args = args
+
+        self.n_heads: int = args.n_heads
+        self.n_kv_heads: int = args.n_kv_heads
+        
+        self.repeats = self.n_heads // self.n_kv_heads
+        self.sliding_window = self.args.sliding_window
+
+        self.scale = self.args.head_dim**-0.5
+
+        self.wq = nn.Linear(
+            args.dim,
+            args.n_heads * args.head_dim,
+            bias=False
+        )
+        self.wk = nn.Linear(
+            args.dim,
+            args.n_kv_heads * args.head_dim,
+            bias=False
+        )
+        self.wv = nn.Linear(
+            args.dim,
+            args.n_kv_heads * args.head_dim,
+            bias=False
+        )
+        self.wo = nn.Linear(
+            args.n_heads * args.head_dim,
+            args.dim,
+            bias=False
+        )
+        
+
+    def forward(
+        self, x: Tensor, 
+        freqs_cis: Tensor,
+        cache: CacheView,
+    ) -> Tensor:
+        seqlen_sum, _ = x.shape
+
+        xq, xk, xv = self.wq(x), self.wk(x), self.wv(x)
+        xq = xq.view(seqlen_sum, self.n_heads, self.args.head_dim)
+        xk = xk.view(seqlen_sum, self.n_kv_heads, self.args.head_dim)
+        xv = xv.view(seqlen_sum, self.n_kv_heads, self.args.head_dim)
+        xq, xk = apply_rotary_emb(xq, xk, freqs_cis=freqs_cis)
+
+        if cache.prefill:
+            key, val = cache.interleave_kv(xk, xv)
+            cache.update(xk, xv)
+        else: 
+            cache.update(xk, xv)
+            key, val = cache.key, cache.value
+            key = key.view(seqlen_sum * cache.sliding_window, self.n_kv_heads, self.args.head_dim)
+            val = val.view(seqlen_sum * cache.sliding_window, self.n_kv_heads, self.args.head_dim)
+
+        # Repeat keys and values to match number of query heads
+        key, val = repeat_kv(key, val, self.repeats, dim=1)
+
+        # xformers requires (B=1, S, H, D)
+        xq, key, val = xq[None, ...], key[None, ...], val[None, ...]
+        output = memory_efficient_attention(xq, key, val, cache.mask)
+
+        return self.wo(output.view_as(x))
