+++ mistral/layers/feed_forward.py
@@ -1,0 +1,26 @@
+from mistral.config.model_config import ModelArgs
+from torch import nn
+from torch import Tensor
+
+class FeedForward(nn.Module):
+    def __init__(self, args: ModelArgs):
+        super().__init__()
+
+        self.w1 = nn.Linear(
+            args.dim,
+            args.hidden_dim,
+            bias=False
+        )
+        self.w2 = nn.Linear(
+            args.hidden_dim,
+            args.dim,
+            bias=False
+        )
+        self.w3 = nn.Linear(
+            args.dim,
+            args.hidden_dim,
+            bias=False
+        )
+
+    def forward(self, x) -> Tensor:
+        return self.w2(nn.functional.silu(self.w1(x)) * self.w3(x))
