+++ mistral/model.py
@@ -5,138 +5,12 @@ from pathlib import Path
 import json
 from typing import List
 
-from mistral.rope import precompute_freqs_cis, apply_rotary_emb
+from mistral.rope import precompute_freqs_cis
 from mistral.cache import CacheView, RotatingBufferCache
-
-from xformers.ops.fmha import (
-    memory_efficient_attention,
-)
-
-
-@dataclass
-class ModelArgs:
-    dim: int
-    n_layers: int
-    head_dim: int
-    hidden_dim: int
-    n_heads: int
-    n_kv_heads: int
-    sliding_window: int
-    norm_eps: float
-    vocab_size: int
-
-    max_batch_size: int = 0
-
-
-def repeat_kv(keys: torch.Tensor, values: torch.Tensor, repeats: int, dim: int):
-    keys = torch.repeat_interleave(keys, repeats=repeats, dim=dim)
-    values = torch.repeat_interleave(values, repeats=repeats, dim=dim)
-    return keys, values
-
-
-class Attention(nn.Module):
-    def __init__(self, args: ModelArgs):
-        super().__init__()
-        self.args = args
-
-        self.n_heads: int = args.n_heads
-        self.n_kv_heads: int = args.n_kv_heads
-        
-        self.repeats = self.n_heads // self.n_kv_heads
-        self.sliding_window = self.args.sliding_window
-
-        self.scale = self.args.head_dim**-0.5
-
-        self.wq = nn.Linear(
-            args.dim,
-            args.n_heads * args.head_dim,
-            bias=False
-        )
-        self.wk = nn.Linear(
-            args.dim,
-            args.n_kv_heads * args.head_dim,
-            bias=False
-        )
-        self.wv = nn.Linear(
-            args.dim,
-            args.n_kv_heads * args.head_dim,
-            bias=False
-        )
-        self.wo = nn.Linear(
-            args.n_heads * args.head_dim,
-            args.dim,
-            bias=False
-        )
-        
-
-    def forward(
-        self, x: torch.Tensor, 
-        freqs_cis: torch.Tensor,
-        cache: CacheView,
-    ) -> torch.Tensor:
-        seqlen_sum, _ = x.shape
-
-        xq, xk, xv = self.wq(x), self.wk(x), self.wv(x)
-        xq = xq.view(seqlen_sum, self.n_heads, self.args.head_dim)
-        xk = xk.view(seqlen_sum, self.n_kv_heads, self.args.head_dim)
-        xv = xv.view(seqlen_sum, self.n_kv_heads, self.args.head_dim)
-        xq, xk = apply_rotary_emb(xq, xk, freqs_cis=freqs_cis)
-
-        if cache.prefill:
-            key, val = cache.interleave_kv(xk, xv)
-            cache.update(xk, xv)
-        else: 
-            cache.update(xk, xv)
-            key, val = cache.key, cache.value
-            key = key.view(seqlen_sum * cache.sliding_window, self.n_kv_heads, self.args.head_dim)
-            val = val.view(seqlen_sum * cache.sliding_window, self.n_kv_heads, self.args.head_dim)
-
-        # Repeat keys and values to match number of query heads
-        key, val = repeat_kv(key, val, self.repeats, dim=1)
-
-        # xformers requires (B=1, S, H, D)
-        xq, key, val = xq[None, ...], key[None, ...], val[None, ...]
-        output = memory_efficient_attention(xq, key, val, cache.mask)
-
-        return self.wo(output.view_as(x))
-
-
-class FeedForward(nn.Module):
-    def __init__(self, args: ModelArgs):
-        super().__init__()
-
-        self.w1 = nn.Linear(
-            args.dim,
-            args.hidden_dim,
-            bias=False
-        )
-        self.w2 = nn.Linear(
-            args.hidden_dim,
-            args.dim,
-            bias=False
-        )
-        self.w3 = nn.Linear(
-            args.dim,
-            args.hidden_dim,
-            bias=False
-        )
-
-    def forward(self, x) -> torch.Tensor:
-        return self.w2(nn.functional.silu(self.w1(x)) * self.w3(x))
-
-
-class RMSNorm(torch.nn.Module):
-    def __init__(self, dim: int, eps: float = 1e-6):
-        super().__init__()
-        self.eps = eps
-        self.weight = nn.Parameter(torch.ones(dim))
-
-    def _norm(self, x):
-        return x * torch.rsqrt(x.pow(2).mean(-1, keepdim=True) + self.eps)
-
-    def forward(self, x):
-        output = self._norm(x.float()).type_as(x)
-        return output * self.weight
+from mistral.layers.attention import Attention
+from mistral.layers.feed_forward import FeedForward
+from mistral.layers.rms_norm import RMSNorm
+from mistral.config.model_config import ModelArgs
 
 
 class TransformerBlock(nn.Module):
